reformatting and editing changes
